Conflicts:
	sdpm-project/business-impl/src/main/java/org/tinygroup/sdpm/project/biz/impl/TaskManagerImpl.java
	sdpm-project/business-interface/src/main/java/org/tinygroup/sdpm/project/biz/inter/TaskManager.java
	sdpm-project/dao-impl/src/main/java/org/tinygroup/sdpm/project/dao/impl/ProjectTaskDaoImpl.java
	sdpm-project/service-impl/src/main/java/org/tinygroup/sdpm/project/service/impl/TaskServiceImpl.java
	sdpm-web/src/main/java/org/tinygroup/sdpm/action/project/TaskAction.java
	sdpm-web/src/main/java/org/tinygroup/sdpm/action/project/util/TaskStatusUtil.java

+++ sdpm-project/business-impl/src/main/java/org/tinygroup/sdpm/project/biz/impl/ProjectProductManagerImpl.java
@@ -19,18 +19,19 @@ public class ProjectProductManagerImpl implements ProjectProductManager {
     @Autowired
     private ProjectProductDao projectProductDao;
 
-    public void addLink(Integer[] array, Integer projectId) {
-        projectProductDao.deleteByProjectId(projectId);
+    public void addLink(Integer[] productArray, Integer projectId) {
+        if (productArray.length < 1) {
+            return;
+        }
         List<ProjectProduct> list = new ArrayList<ProjectProduct>();
-        for (Integer productId : array) {
+        for (Integer productId : productArray) {
             ProjectProduct t = new ProjectProduct();
             t.setProjectId(projectId);
             t.setProductId(productId);
             list.add(t);
         }
-        if (list != null || !list.isEmpty()) {
-            projectProductDao.batchInsert(list);
-        }
+        projectProductDao.deleteByProjectId(projectId);
+        projectProductDao.batchInsert(list);
     }
 
     public List<ProjectProduct> findList(ProjectProduct projectProduct) {
